Feature #70119, ValueRangeController service proxy (#634)
